+++ modules/face/instantid.py
@@ -71,8 +71,8 @@ def instant_id(p: processing.StableDiffusionProcessing, app, source_images, stre
     p.task_args['controlnet_conditioning_scale'] = float(conditioning)
     p.task_args['ip_adapter_scale'] = float(strength)
     shared.log.debug(f"InstantID args: {p.task_args}")
-    p.task_args['prompt'] = p.all_prompts[0] # override all logic
-    p.task_args['negative_prompt'] = p.all_negative_prompts[0]
+    p.task_args['prompt'] = p.all_prompts[0] if p.all_prompts is not None else p.prompt
+    p.task_args['negative_prompt'] = p.all_negative_prompts[0] if p.all_negative_prompts is not None else p.negative_prompt
     p.task_args['image_embeds'] = face_embeds[0] # overwrite placeholder
 
     # run processing