+++ modules/ui_gallery.py
@@ -1,10 +1,10 @@
 import os
 from datetime import datetime
+from urllib.parse import unquote
 import gradio as gr
 from PIL import Image
 from modules import shared, ui_symbols, ui_common, images, ui_control_helpers
 from modules.ui_components import ToolButton
-from urllib.parse import unquote
 
 def read_media(fn):
     fn = unquote(fn).replace('%3A', ':')